+++ toolkit/mozapps/update/common/pathhash.cpp
@@ -6,19 +6,16 @@
 #include <wincrypt.h>
 #include "pathhash.h"
 
-
 /**
  * Converts a binary sequence into a hex string
  *
  * @param hash      The binary data sequence
  * @param hashSize  The size of the binary data sequence
- * @param hexString A buffer to store the hex string, must be of 
+ * @param hexString A buffer to store the hex string, must be of
  *                  size 2 * @hashSize
-*/
-static void
-BinaryDataToHexString(const BYTE *hash, DWORD &hashSize, 
-                      LPWSTR hexString)
-{
+ */
+static void BinaryDataToHexString(const BYTE *hash, DWORD &hashSize,
+                                  LPWSTR hexString) {
   WCHAR *p = hexString;
   for (DWORD i = 0; i < hashSize; ++i) {
     wsprintfW(p, L"%.2x", hash[i]);
@@ -34,22 +31,20 @@ BinaryDataToHexString(const BYTE *hash, DWORD &hashSize,
  * @param  hash     Output buffer to store hash, must be freed by the caller
  * @param  hashSize The number of bytes in the output buffer
  * @return TRUE on success
-*/
-static BOOL
-CalculateMD5(const char *data, DWORD dataSize, 
-             BYTE **hash, DWORD &hashSize)
-{
+ */
+static BOOL CalculateMD5(const char *data, DWORD dataSize, BYTE **hash,
+                         DWORD &hashSize) {
   HCRYPTPROV hProv = 0;
   HCRYPTHASH hHash = 0;
 
   if (!CryptAcquireContext(&hProv, nullptr, nullptr, PROV_RSA_FULL,
                            CRYPT_VERIFYCONTEXT)) {
-    if (NTE_BAD_KEYSET != GetLastError()) {
+    if ((DWORD)NTE_BAD_KEYSET != GetLastError()) {
       return FALSE;
     }
- 
+
     // Maybe it doesn't exist, try to create it.
-    if (!CryptAcquireContext(&hProv, nullptr, nullptr, PROV_RSA_FULL, 
+    if (!CryptAcquireContext(&hProv, nullptr, nullptr, PROV_RSA_FULL,
                              CRYPT_VERIFYCONTEXT | CRYPT_NEWKEYSET)) {
       return FALSE;
     }
@@ -59,17 +54,16 @@ CalculateMD5(const char *data, DWORD dataSize,
     return FALSE;
   }
 
-  if (!CryptHashData(hHash, reinterpret_cast<const BYTE*>(data), 
-                    dataSize, 0)) {
+  if (!CryptHashData(hHash, reinterpret_cast<const BYTE *>(data), dataSize,
+                     0)) {
     return FALSE;
   }
 
   DWORD dwCount = sizeof(DWORD);
-  if (!CryptGetHashParam(hHash, HP_HASHSIZE, (BYTE *)&hashSize, 
-                        &dwCount, 0)) {
+  if (!CryptGetHashParam(hHash, HP_HASHSIZE, (BYTE *)&hashSize, &dwCount, 0)) {
     return FALSE;
   }
-  
+
   *hash = new BYTE[hashSize];
   ZeroMemory(*hash, hashSize);
   if (!CryptGetHashParam(hHash, HP_HASHVAL, *hash, &hashSize, 0)) {
@@ -81,7 +75,7 @@ CalculateMD5(const char *data, DWORD dataSize,
   }
 
   if (hProv) {
-    CryptReleaseContext(hProv,0);
+    CryptReleaseContext(hProv, 0);
   }
 
   return TRUE;
@@ -91,22 +85,19 @@ CalculateMD5(const char *data, DWORD dataSize,
  * Converts a file path into a unique registry location for cert storage
  *
  * @param  filePath     The input file path to get a registry path from
- * @param  registryPath A buffer to write the registry path to, must 
+ * @param  registryPath A buffer to write the registry path to, must
  *                      be of size in WCHARs MAX_PATH + 1
  * @return TRUE if successful
-*/
-BOOL
-CalculateRegistryPathFromFilePath(const LPCWSTR filePath, 
-                                  LPWSTR registryPath)
-{
-  size_t filePathLen = wcslen(filePath); 
+ */
+BOOL CalculateRegistryPathFromFilePath(const LPCWSTR filePath,
+                                       LPWSTR registryPath) {
+  size_t filePathLen = wcslen(filePath);
   if (!filePathLen) {
     return FALSE;
   }
 
   // If the file path ends in a slash, ignore that character
-  if (filePath[filePathLen -1] == L'\\' || 
-      filePath[filePathLen - 1] == L'/') {
+  if (filePath[filePathLen - 1] == L'\\' || filePath[filePathLen - 1] == L'/') {
     filePathLen--;
   }
 
@@ -121,19 +112,18 @@ CalculateRegistryPathFromFilePath(const LPCWSTR filePath,
 
   BYTE *hash;
   DWORD hashSize = 0;
-  if (!CalculateMD5(reinterpret_cast<const char*>(lowercasePath), 
-                    filePathLen * 2, 
-                    &hash, hashSize)) {
+  if (!CalculateMD5(reinterpret_cast<const char *>(lowercasePath),
+                    filePathLen * 2, &hash, hashSize)) {
     delete[] lowercasePath;
     return FALSE;
   }
   delete[] lowercasePath;
 
-  LPCWSTR baseRegPath = L"SOFTWARE\\Mozilla\\"
-    L"MaintenanceService\\";
+  LPCWSTR baseRegPath =
+      L"SOFTWARE\\Mozilla\\"
+      L"MaintenanceService\\";
   wcsncpy(registryPath, baseRegPath, MAX_PATH);
-  BinaryDataToHexString(hash, hashSize, 
-                        registryPath + wcslen(baseRegPath));
+  BinaryDataToHexString(hash, hashSize, registryPath + wcslen(baseRegPath));
   delete[] hash;
   return TRUE;
 }
